feat(dotnet-sdk): Upgrade to C# implementation (#2453)
* chore(dotnet-sdk): replace the wrapper to C# implementation

* chore(dotnet-sdk): change option structure

* chore(dotnet-sdk): bump version
