--- packages/dotnet-sdk/src/TeamsFx/Model/LogLevel.cs
@@ -1,28 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-namespace Microsoft.TeamsFx.Model
-{
-    /// <summary>
-    /// Log Level
-    /// </summary>
-    public enum LogLevel
-    {
-        /// <summary>
-        /// Show verbose, information, warning and error message.
-        /// </summary>
-        Verbose = 0,
-        /// <summary>
-        /// Show information, warning and error message.
-        /// </summary>
-        Info,
-        /// <summary>
-        /// Show warning and error message.
-        /// </summary>
-        Warn,
-        /// <summary>
-        /// Show error message.
-        /// </summary>
-        Error
-    }
-}
