+++ packages/dotnet-sdk/src/TeamsFx/wwwroot/jsInterop.js
@@ -1,23 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { TeamsUserCredential, setLogFunction } from "./teamsfx.js"
-
-// create class instance
-export function createTeamsUserCredential() {
-    return new TeamsUserCredential();
-}
-
-// call .NET function
-export function setLogFunctionCallback(logFunctionCallbackRef) {
-    setLogFunction((level, message) => {
-        logFunctionCallbackRef.invokeMethodAsync('Invoke', level, message);
-    });
-}
-
-export function clearLogFunctionCallback() {
-    setLogFunction(undefined);
-}
 
 export function initialize() {
     const initializeTeamsSdkTimeoutInMillisecond = 5000;
@@ -73,5 +56,3 @@ export function getAuthToken() {
         }
     });
 }
-
-export * from "./teamsfx.js"
